Version: Initial add for version command
